Separate command for copy_env and more refactoring
